akela-backend/apps/prescriptions/serializers.py
from rest_framework import serializers
from .models import Prescription


class PrescriptionSerializer(serializers.ModelSerializer):
    class Meta:
        model   = Prescription
        exclude = ['user']
