akela-backend/requirements.txt
Django==4.2
djangorestframework==3.14
django-cors-headers==4.3
python-dotenv==1.0
gunicorn==21.2
psycopg2-binary==2.9
dj-database-url==2.1
djangorestframework-simplejwt==5.3
django-filter==23.3
drf-spectacular==0.27
razorpay
stripe
cloudinary==1.36
django-cloudinary-storage==0.3
Pillow==10.1
resend==0.7
django-redis==5.4
redis==5.0
celery==5.3
whitenoise==6.6
opencv-python-headless==4.9.0.80
mediapipe==0.10.9
numpy>=1.23,<2.0
